requirements.txt
discord.py
aiosqlite
python-dotenv
quart
httpx
aiosmtplib
google-api-python-client